Start refactor of game mode - Hopefully opens up possibility of creating new HeadlessGame - TODO: Remove CoopGame.Instance - Not tested yet
